Refactor + AGENTS file
