+++ README.md
@@ -115,7 +115,7 @@ build.bat
 
 ### Manual Build
 ```batch
-cl.exe src\oled_aegis.c /Fe:oled_aegis.exe /O2 /MD /link user32.lib shell32.lib ole32.lib uuid.lib gdi32.lib advapi32.lib comctl32.lib powrprof.lib psapi.lib dwmapi.lib
+cl.exe src\oled_aegis.c src\util.c src\logging.c src\config.c src\monitors.c src\media.c src\screensaver.c src\settings.c /Fe:oled_aegis.exe /O2 /MD /link user32.lib shell32.lib ole32.lib uuid.lib gdi32.lib advapi32.lib comctl32.lib powrprof.lib psapi.lib dwmapi.lib
 ```
 
 See [BUILD.md](BUILD.md) for more information.
